+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Greatly expanded Test Kitchen coverage and platform support.
 - added more asserts to the helper for vhosts, users, etc.
+- [COOK-2397]: Added upstart support to rabbitmq cookbook
 - [COOK-2830]: Use a notify for server restart, instead of defining a new service
 
 ### Bug